codex: move app-server attestation into module
